Added enums. Refactoring.

+++ src/main/java/net/usrlib/cms/data/CourseAssignmentsData.java
@@ -30,9 +30,9 @@ public class CourseAssignmentsData extends CsvDataLoader {
 			if (parts.length > 0) {
 				try {
 					preparedStatement = connection.prepareStatement(CourseAssignmentsTable.INSERT_SQL);
-					preparedStatement.setLong(1, Long.valueOf(parts[0]));
-					preparedStatement.setLong(2, Long.valueOf(parts[1]));
-					preparedStatement.setLong(3, Long.valueOf(parts[2]));
+					preparedStatement.setInt(1, Integer.valueOf(parts[0]));
+					preparedStatement.setInt(2, Integer.valueOf(parts[1]));
+					preparedStatement.setInt(3, Integer.valueOf(parts[2]));
 
 					preparedStatement.execute();
 				} catch (SQLException e) {
